requirements.txt
gitpython
faiss-cpu
sentence-transformers
tree-sitter
numpy
pandas
tqdm